tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-fill-mode_default.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-fill-mode default</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-fill-mode-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="animation-fill-mode does not inherit by
   default">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes move {
         0% {
            left: 100px;
         }
         100% {
            left: 100px;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: gray;
         position: relative;
         animation-duration: 0.5s;
         animation-delay: 0.5s;
      }
      div > div {
         background: blue;
      }
      .test {
         animation-fill-mode: both;
      }
      .test, .test > div {
         animation-name: move;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var divs = document.querySelectorAll('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var outerStyle = getComputedStyle(divs[0], null);
         var innerStyle = getComputedStyle(divs[1], null);
         var passed = false;
         setTimeout(function() {
            passed = (outerStyle.left == '100px' && innerStyle.left == 'auto');
         }, 250);
         setTimeout(function() {
            passed = (passed && outerStyle.left == '100px'
                      && innerStyle.left == '100px');
         }, 750);
         setTimeout(function() {
            var fill = innerStyle.getPropertyValue('animation-fill-mode');
            passed = (passed && outerStyle.left == '100px' && fill == 'none'
                      && innerStyle.left == 'auto');
            test(function(){assert_true(passed);}, testName); done();
            log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
         }, 1250);
         divs[0].className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>There should be a blue square below, and after a short delay it should
   jump to the right, uncovering a gray square, then jump back on top of the
   gray square again.</p>
   <p>Testing...</p>
   <div>
      <div></div>
   </div>
</body>
</html>
